Fixed up cloak (at long, long last)

+++ src/gfx/halo_system.h
@@ -33,7 +33,7 @@ class HaloSystem {
   }
   unsigned int  AddHalo (const char * filename, const QVector &loc, const Vector &size, const GFXColor & col, std::string halo_type/*when it grows*/, float activation_speed);
   bool ShouldDraw (float speedsquared);
-  void Draw (const Matrix & trans, const Vector & scale, short halo_alpha, float nebdist, float hullpercentage, const Vector & velocity, int faction);//draws smoke and damage fx
+  void Draw (const Matrix & trans, const Vector & scale, int halo_alpha, float nebdist, float hullpercentage, const Vector & velocity, int faction);//draws smoke and damage fx
   //  unsigned int size(); {return halo.size();}
   void SetPosition (unsigned int which, const QVector &loc);
   void SetSize (unsigned int which, const Vector &scale);
